[Fix #7085] Fix issue in Layout/IndentHeredoc
When a squiggly heredoc is given on the second (or higher) line of a method
argument, `Layout/IndentHeredoc` wanted me to indent the heredoc two spaces
more than the last line of method arguments.

With this commit, a heredoc given as argument (to a `send` type node) will be
indented two spaces (configurable) more than the send node.